1. OPEN_MACRO* заменены на OPEN_FILEPANEL. тип вызова передаётся в нулевом параметре с использованием MACROCALLTYPE. 2. MCTL_ISFARSPRING и MCTL_CALLFAR убраны. вместо них в PluginStartupInfo устанавливается Private ссылкой на MacroPrivateInfo. проверка на nullptr Private аналогична MCTL_ISFARSPRING. вызов MacroPrivateInfo.CallFar аналогичен MCTL_CALLFAR.

+++ far/macro.hpp
@@ -244,7 +244,7 @@ class KeyMacro
 		// Поместить временное строковое представление макроса
 		bool PostNewMacro(const wchar_t *PlainText,UINT64 Flags=0,DWORD AKey=0,bool onlyCheck=false);
 		bool ParseMacroString(const wchar_t *Sequence,bool onlyCheck=false);
-		int CallFar(int OpCode, FarMacroCall* Data);
+		intptr_t CallFar(intptr_t OpCode, FarMacroCall* Data);
 };
 
 const wchar_t *eStackAsString(int Pos=0);
